Make expiry U256

+++ contracts/interfaces/IAVNBridge.sol
@@ -30,15 +30,15 @@ interface IAVNBridge {
   function updateLowerCall(bytes2 callId, uint256 numBytes) external;
 
   // Owner or validators only
-  function triggerGrowth(uint128 amount, uint32 period, uint64 expiry, uint32 t2TransactionId, bytes calldata confirmations)
+  function triggerGrowth(uint128 amount, uint32 period, uint256 expiry, uint32 t2TransactionId, bytes calldata confirmations)
       external;
 
   // Validators only
-  function registerValidator(bytes calldata t1PublicKey, bytes32 t2PublicKey, uint64 expiry, uint32 t2TransactionId,
+  function registerValidator(bytes calldata t1PublicKey, bytes32 t2PublicKey, uint256 expiry, uint32 t2TransactionId,
       bytes calldata confirmations) external;
-  function deregisterValidator(bytes calldata t1PublicKey, bytes32 t2PublicKey, uint64 expiry, uint32 t2TransactionId,
+  function deregisterValidator(bytes calldata t1PublicKey, bytes32 t2PublicKey, uint256 expiry, uint32 t2TransactionId,
       bytes calldata confirmations) external;
-  function publishRoot(bytes32 rootHash, uint64 expiry, uint32 t2TransactionId, bytes calldata confirmations) external;
+  function publishRoot(bytes32 rootHash, uint256 expiry, uint32 t2TransactionId, bytes calldata confirmations) external;
 
   // Public
   function releaseGrowth(uint32 period) external;
